roll rebuild up into reify
